# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI股票助手API",
    "description": "AI股票助手后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "AI股票助手开发团队"
    }
  },
  "host": "localhost:3033",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/aisignals/create": {
      "post": {
        "tags": ["AI信号"],
        "summary": "创建AI信号记录",
        "description": "创建单个AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAiSignalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/create-batch": {
      "post": {
        "tags": ["AI信号"],
        "summary": "批量创建AI信号记录",
        "description": "批量创建多个AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchCreateAiSignalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "批量创建成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/aisignals/get-by-symbol": {
      "post": {
        "tags": ["AI信号"],
        "summary": "根据股票代码获取AI信号记录",
        "description": "根据股票代码获取相关AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalBySymbolRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/get-by-signal-type": {
      "post": {
        "tags": ["AI信号"],
        "summary": "根据信号类型获取AI信号记录",
        "description": "根据买入、卖出或持有信号类型获取AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalByTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/aisignals/get-by-symbol-and-time-range": {
      "post": {
        "tags": ["AI信号"],
        "summary": "根据股票代码和时间范围获取AI信号记录",
        "description": "根据股票代码和时间范围获取相关AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalBySymbolAndTimeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/get-by-confidence-range": {
      "post": {
        "tags": ["AI信号"],
        "summary": "根据置信度范围获取AI信号记录",
        "description": "根据置信度范围获取AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalByConfidenceRangeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/get-latest": {
      "post": {
        "tags": ["AI信号"],
        "summary": "获取最新AI信号记录",
        "description": "获取最新的AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetLatestAiSignalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/get-by-symbol-and-type": {
      "post": {
        "tags": ["AI信号"],
        "summary": "根据股票代码和信号类型获取记录",
        "description": "根据股票代码和信号类型获取AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalBySymbolAndTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/get-high-confidence": {
      "post": {
        "tags": ["AI信号"],
        "summary": "获取高置信度信号",
        "description": "获取高置信度的AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetHighConfidenceSignalsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/stats": {
      "post": {
        "tags": ["AI信号"],
        "summary": "获取AI信号统计信息",
        "description": "获取AI信号相关的统计信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAiSignalStatsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/update": {
      "post": {
        "tags": ["AI信号"],
        "summary": "更新AI信号记录",
        "description": "更新指定的AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAiSignalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/delete": {
      "post": {
        "tags": ["AI信号"],
        "summary": "删除AI信号记录",
        "description": "删除指定的AI信号记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteAiSignalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/aisignals/clean-old-data": {
      "post": {
        "tags": ["AI信号"],
        "summary": "清理过期数据",
        "description": "清理指定天数之前的过期AI信号数据",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CleanOldDataRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "清理成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "description": "返回的数据"
        },
        "message": {
          "type": "string",
          "description": "操作成功/报错信息"
        },
        "result": {
          "type": "integer",
          "description": "100表示接口正常，其他数字表示错误码"
        },
        "success": {
          "type": "boolean",
          "description": "true表示请求成功 false表示接口报错"
        }
      }
    },
...
    "CleanOldDataRequest": {
      "type": "object",
      "properties": {
        "daysToKeep": {
          "type": "integer",
          "description": "保留天数"
        }
      }
    },
    "CreateAiSignalRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "signalTime": {
          "type": "string",
          "format": "date-time",
          "description": "信号时间"
        },
        "signalType": {
          "type": "string",
          "enum": ["buy", "sell", "hold"],
          "description": "AI信号类型"
        },
        "confidence": {
          "type": "integer",
          "description": "信号置信度(0-100)"
        },
        "modelVersion": {
          "type": "string",
          "description": "AI模型版本号"
        },
        "description": {
          "type": "string",
          "description": "信号说明或模型解释"
        }
      }
    },
    "BatchCreateAiSignalRequest": {
      "type": "object",
      "required": ["aiSignalDataList"],
      "properties": {
        "aiSignalDataList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CreateAiSignalRequest"
          }
        }
      }
    },
    "GetAiSignalBySymbolRequest": {
      "type": "object",
      "required": ["symbol"],
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        }
      }
    },
    "GetAiSignalByTypeRequest": {
      "type": "object",
      "required": ["signalType"],
      "properties": {
        "signalType": {
          "type": "string",
          "enum": ["buy", "sell", "hold"],
          "description": "信号类型"
        }
      }
    },
...
    "GetAiSignalBySymbolAndTimeRequest": {
      "type": "object",
      "required": ["symbol", "startTime", "endTime"],
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "startTime": {
          "type": "string",
          "format": "date-time",
          "description": "开始时间"
        },
        "endTime": {
          "type": "string",
          "format": "date-time",
          "description": "结束时间"
        }
      }
    },
    "GetAiSignalByConfidenceRangeRequest": {
      "type": "object",
      "required": ["minConfidence", "maxConfidence"],
      "properties": {
        "minConfidence": {
          "type": "integer",
          "description": "最小置信度"
        },
        "maxConfidence": {
          "type": "integer",
          "description": "最大置信度"
        }
      }
    },
    "GetLatestAiSignalRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "limit": {
          "type": "integer",
          "description": "限制数量"
        }
      }
    },
    "GetAiSignalBySymbolAndTypeRequest": {
      "type": "object",
      "required": ["symbol", "signalType"],
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "signalType": {
          "type": "string",
          "enum": ["buy", "sell", "hold"],
          "description": "信号类型"
        }
      }
    },
    "GetHighConfidenceSignalsRequest": {
      "type": "object",
      "properties": {
        "minConfidence": {
          "type": "integer",
          "description": "最小置信度"
        },
        "limit": {
          "type": "integer",
          "description": "限制数量"
        }
      }
    },
    "GetAiSignalStatsRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "startTime": {
          "type": "string",
          "format": "date-time",
          "description": "开始时间"
        },
        "endTime": {
          "type": "string",
          "format": "date-time",
          "description": "结束时间"
        }
      }
    },
    "UpdateAiSignalRequest": {
      "type": "object",
      "required": ["id", "updateData"],
      "properties": {
        "id": {
          "type": "integer",
          "description": "AI信号ID"
        },
        "updateData": {
          "$ref": "#/definitions/CreateAiSignalRequest",
          "description": "更新数据"
        }
      }
    },
    "DeleteAiSignalRequest": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "integer",
          "description": "AI信号ID"
        }
      }
    }
  },
...
}